% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatplots.R
\name{axis_units}
\alias{axis_units}
\title{Change Axis Scale Units}
\usage{
axis_units(x)
}
\arguments{
\item{x}{value in meters}
}
\description{
Divides the scale of the axis by 1,000 so units show up in kilometers instead of meters.
}
\examples{
## Convert 1,200 meters to 1.2 KM
scale_axis(1200)
}
\keyword{axis}
\keyword{scale,}
