% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatplots.R
\name{change_gradient_breaks}
\alias{change_gradient_breaks}
\title{Modify Color Ramp and Breaks}
\usage{
change_gradient_breaks(new_colors = c("blue", "green", "yellow", "orange",
  "red"), mapped_break_values = c(0, 0.25, 0.5, 0.75, 1))
}
\arguments{
\item{new_colors}{vector of colors to apply to the heatplot fill}

\item{mapped_break_values}{vector of the break values to use for the specific colors (length of vector must match new_colors)}
}
\description{
Modify default colors and breaks used in the heatplot. The list of the colors and the list of breaks values must be equal.
The default color ramp is blue-green-yellow-orange-red with breaks at 0%, 25%, 50%, 75%, 100% of the water quality variable.
The break values can either be percentages (make sure the break values start with zero and end with 1) or the actual values
of the variable. The actual values will be remapped internally to be between zero and one.
}
\keyword{breaks}
\keyword{color,}
