% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatplots.R
\name{heatplot}
\alias{heatplot}
\title{Heatplot}
\usage{
heatplot(df, dateField, distanceField, wqVariable, title,
  formating = heatplot.format.default)
}
\arguments{
\item{df}{dataframe (from Export Heatplot data csv)}

\item{dateField}{field name with the data time information for the water quality point}

\item{distanceField}{field name with the distance information for the water quality point}

\item{title}{the title for the plot}

\item{formating}{ggplot formating options to alter plot appearance}

\item{wqVaariable}{field name for the water quality variable to symbolize}
}
\description{
Plots month by river distance symbolized by water quality variable
}
\examples{
## Make a heatplot
p <- heatplot(ex_data, "date_time", "m_value", "temp", "Example Reach Temperature", heatplot.format.default)
}
\keyword{heatplot}
\keyword{plot,}
