% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatplots.R
\name{month_startdate}
\alias{month_startdate}
\title{Start of month}
\usage{
month_startdate(date)
}
\arguments{
\item{date}{Date Object}
}
\description{
This function gets the first of the month for given date_time object
}
\examples{
month_startdate()
}
\keyword{date}
\keyword{month}
\keyword{start}
