% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatplots.R
\name{unique_date_breaks}
\alias{unique_date_breaks}
\title{Date Breaks}
\usage{
unique_date_breaks(dataframe, dateField)
}
\arguments{
\item{dataframe}{R dataframe object with water quality data.}

\item{dateField}{Name of the field with dates}
}
\description{
This function gets the unqiue months and returns the middle of the month (the 15th) to use as the date breaks
}
\examples{
unique_date_breaks()
}
\keyword{date_breaks}
