% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatplots.R
\name{heatplot.format.default}
\alias{heatplot.format.default}
\title{Default Format for Heatplot}
\usage{
heatplot.format.default(p)
}
\description{
The defualt formating options for the heatplot
}
\keyword{heatplot}
\keyword{plot,}
